Refactoring rules tests
