--- src/lint/rules/test/indentation/mod.rs
@@ -1,1 +1,0 @@
-pub mod in5;
