+++ src/lint/rules/tests/mod.rs
@@ -1,0 +1,3 @@
+pub mod indentation;
+pub mod spacing;
+pub mod common;
